1.首页增加toggle 2.我的信箱 3.其它bug修复 4.退出登录

+++ src/main/java/com/xkf/cashbook/mysql/model/FamilyApplyDO.java
@@ -31,7 +31,7 @@ public class FamilyApplyDO {
 
     private Integer applyStatus;
 
-    private Integer applyUserId;
+    private Long applyUserId;
 
     private LocalDateTime applyTime;
